requirements.txt
asgiref==3.4.1
Django==3.2.5
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.28
pytz==2021.1
sqlparse==0.4.1
tinydb==4.5.1
